chore(core): reusable list helper
Move the list string formatting helper from the legend module into the text module in order to position it for more regular reuse with other text content.
